Implement mprotect for sel4 rumprun
Mprotect is used to implement guard pages by pthreads, and noone likes
running off the stack.

As part of this change the default page size is set to the rump system
page size such that pages can be unmapped and remapped without being broken up and
retyped. Previously the page size was the largest for the amount of
memory given to rump.

Rumprun will still use large pages if CONFIG_USE_LARGE_PAGES is set,
and mprotect is not implemented.