Create init-sources.sh for patch files
This moves the functionality for checking out the rumprun submodules and
applying our patch files from a Makefile into a script. Originally,
modifying our source directories in our build scripts is a bit of a
smell and when moving to cmake it should be fixed.  Adding the
invocation to Kbuild is for backwards compatibility with the current
way rumprun is checked out and built.